fix: carry money full-precision, round only at export; show 2dp in UI
Member cash had per-trade rounding drift because every money write went
through _int (round-half-up to whole units). Buy+sell at the same price
lost <1 unit each round-trip.

Backend: balance/debt/FD principal+payout/txn amount are now full-precision
floats (REAL columns), no intermediate rounding. Trade cost = price*shares
exact. The ONE rounding point is core.networth.member_amount (the export /
scoreboard boundary), which still rounds half-up to whole units via _int.

Frontend: full-precision floats surfaced as 3 decimals on the member/teller
pages (their fmt had no maximumFractionDigits). Centralize two formatters in
common.js — money() = always 2dp with padded trailing zeros (1.1 -> 1.10),
count() = integer share counts — and route all pages through them.

Tests: 5 assertions re-expressed as pytest.approx of the unchanged formula
(only the dropped round differs). 70 pass.

Note: re-provision an existing DB (setup_db.py --reset) to adopt REAL columns.

+++ app/core/db.py
@@ -3,11 +3,11 @@ import sqlite3
 SCHEMA = """
 CREATE TABLE IF NOT EXISTS members(
   member_id TEXT PRIMARY KEY, pin TEXT UNIQUE NOT NULL,
-  balance INTEGER NOT NULL, balance_accrued_at REAL NOT NULL,
-  debt INTEGER NOT NULL DEFAULT 0, loan_taken_at REAL,
+  balance REAL NOT NULL, balance_accrued_at REAL NOT NULL,
+  debt REAL NOT NULL DEFAULT 0, loan_taken_at REAL,
   relief_claimed INTEGER NOT NULL DEFAULT 0, last_teller_visit_at REAL);
 CREATE TABLE IF NOT EXISTS fixed_deposits(
-  fd_id TEXT PRIMARY KEY, member_id TEXT NOT NULL, principal INTEGER NOT NULL,
+  fd_id TEXT PRIMARY KEY, member_id TEXT NOT NULL, principal REAL NOT NULL,
   term_minutes INTEGER NOT NULL, rate_per_min REAL NOT NULL, created_at REAL NOT NULL,
   matured INTEGER NOT NULL DEFAULT 0, closed INTEGER NOT NULL DEFAULT 0);
 CREATE TABLE IF NOT EXISTS holdings(
@@ -25,7 +25,7 @@ CREATE TABLE IF NOT EXISTS trades(
   side TEXT, shares INTEGER, price REAL, ts REAL, actor TEXT);
 CREATE TABLE IF NOT EXISTS transactions(
   id INTEGER PRIMARY KEY AUTOINCREMENT, member_id TEXT, type TEXT,
-  amount INTEGER, ts REAL, actor TEXT);
+  amount REAL, ts REAL, actor TEXT);
 CREATE TABLE IF NOT EXISTS events(
   id INTEGER PRIMARY KEY AUTOINCREMENT, at_min REAL, stock_id TEXT, pct REAL,
   duration_min REAL, headline TEXT, fired INTEGER NOT NULL DEFAULT 0);
